pyproject.toml
[project]
name = "django-auth-bot"
version = "0.1.0"
description = "Тестовое"
authors = [
    {name = "nosatiy@icloud.com"}
]

license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "pytelegrambotapi (>=4.27.0,<5.0.0)",
    "dramatiq[rabbitmq,watch] (>=1.17.1,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "django-dramatiq (>=0.13.0,<0.14.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

